Parsing headers, HUGE refactor
Now parsing headers so we can search at any time. This also came with a pretty huge refactor for code cleanup and consolidating some data structures (e.g., introducing SortedArray struct that we can use for both parameters and headers and probably routes in the future). There's still some cleanup to do but this improves readability a ton